feat(android): one token source, offline fixtures, service contracts and a component gallery
Step 4 groundwork of the mobile-experience-design skill, as code:

design/tokens.json is now the single platform token source, generated by
.mobile-work/generate-tokens.py from shared-spec/design-tokens.json plus the existing
Compose tokens. It resolves every upstream alias for both themes, turns alpha modifiers
into ARGB literals, names the functional roles the components need, records the layout
facts from design/DESIGN.md, and measures 27 WCAG contrast pairs (0 below target; the six
upstream alias choices that could not reach their target moved one step and carry a note).

The Compose token objects are aligned with it: HeliosSemanticColor.kt (Paper and Carbon),
HeliosSpacing.kt, a pill radius, material tonal elevation and the 0.5 dp inner stroke,
HeliosMotion durations, curves, entrance delays and the Reduce Motion contract, and a
theme that maps semantic roles to Material roles and provides them through a
CompositionLocal. TokensMatchJsonTest fails the build if the two drift.

core/data/service/** holds ten typed contracts (telemetry, today series, live series,
forecast, insights, connection, location, theme, brand, share) that return Loadable, so a
value cannot reach a screen without its state, and whose failures are the FLW-03 classes
F1-F9 with their own words and recovery action. core/data/fixture/** implements them with
deterministic data and nine scenarios: loading, live, empty, partial, stale, offline,
error, denied permission and long content. Nothing touches the network.

The debug component gallery renders every reused component in every variant and state,
packed into 47 pages that each fit one screen, and every page ends with a marker stripe so
a capture can prove nothing was clipped. It is a debug source set: the release APK
contains no trace of it. The entry point is a composable; core/nav and MainActivity are
untouched.

Snapshot v1 parity is now byte-level: integer fields where the PWA rounds, fc and br
omitted when absent, br omitted for helios, PWA key order. A failed register read is NaN
and renders "No data", never 0.

Verified: :app:assembleDebug and :app:assembleRelease exit 0; 17 JVM tests pass; 47 pages
captured on helios35 with every end marker present; design/captures/gallery-components.png
is the contact sheet. Evidence and open risks are in .mobile-work/STATE.md.

+++ android/app/src/main/java/com/helios/core/domain/model/SnapshotPayload.kt
@@ -3,7 +3,15 @@ package com.helios.core.domain.model
 import kotlinx.serialization.Serializable
 
 /**
- * v1 shared-snapshot payload matched verbatim to PWA SnapshotPayload.
+ * v1 shared-snapshot payload, matched verbatim to PWA `src/services/share.ts`.
+ *
+ * The field types follow what `JSON.stringify` prints, not what is convenient:
+ * `lifeKwh`, `soc`, `selfUse` and every `fc` entry are integers, because the PWA passes
+ * each through `Math.round`. A `Double` holding 18420.0 would serialise as "18420.0"
+ * while the PWA writes "18420", and the two links would stop being byte-identical.
+ *
+ * `fc` and `br` are omitted when absent (the PWA writes `undefined` for `br` on the
+ * helios brand), which the encoder does by not encoding nulls.
  */
 @Serializable
 data class SnapshotPayload(
@@ -12,9 +20,9 @@ data class SnapshotPayload(
     val loc: String,
     val ac: Double,
     val todayKwh: Double,
-    val lifeKwh: Double,
-    val soc: Double,
-    val selfUse: Double,
-    val fc: List<Double>? = null,
+    val lifeKwh: Int,
+    val soc: Int,
+    val selfUse: Int,
+    val fc: List<Int>? = null,
     val br: String? = null
 )
